Bleepin' Windows uses LLP64, not LP64
git-svn-id: svn+ssh://svn.code.sf.net/p/libjpeg-turbo/code/trunk@158 632fc199-4ca6-4c93-a231-07263d6284db

+++ jdhuff.h
@@ -75,7 +75,7 @@ EXTERN(void) jpeg_make_d_derived_tbl
 
 #if __WORDSIZE == 64
 
-typedef long bit_buf_type;	/* type of bit-extraction buffer */
+typedef size_t bit_buf_type;	/* type of bit-extraction buffer */
 #define BIT_BUF_SIZE  64		/* size of buffer in bits */
 
 #else
